refactor: types and scope
