+++ src/pages/Profile/MyVehicle/components/VehicleEdit/types.ts
@@ -1,0 +1,6 @@
+import type maintenanceIcons from '@/assets/maintenance';
+
+export type MaintenanceItemProps = {
+  icon: keyof typeof maintenanceIcons;
+  title: string;
+};
