+++ src/shared/types/NavBar.ts
@@ -1,0 +1,10 @@
+export type Tab = {
+  name: string;
+  path: string;
+  icon: {
+    enabled: string;
+    disabled: string;
+  };
+  label: string;
+  notification?: number;
+};
